requirements.txt
Flask
psycopg[binary]
python-dotenv
python-dateutil
